+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 1.0.16
 
+- Added `auto_reduce` option in `sig_fit*` functions to improve signature fitting.
 - Return cosine similarity for sample profile in `sig_fit()`.
 - Set default strategy in `sig_auto_extract()` to 'optimal'.
 - Supported search reference signature index in `get_sig_cancer_type_index()`.